+++ CHANGELOG.md
@@ -6,10 +6,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Calendar Versioning](https://calver.org/) of
 the following form: YYYY.0M.0D.
 
-## [2024.09.16]
-
+## [2024.10.31]
 ### Added
+- Designed a abstract class named "setlistSecondary" for the ConcertSetlist component
 
+## [2024.10.02]
+### Added
+- Designed a proof of concept for ConcertSetlist component
+- Showed all of the basic methods other than moveSong, findSong, and getSongAt
+
+## [2024.09.16]
+### Added
 - Designed a Concert Setlist component
 - Designed a Playing Cards component
 - Designed a Weather Tracker component